requirement.txt
flask==2.3.3
python-dotenv==0.21.1
twilio==8.5.0
deepgram-sdk==2.11.0
openai==0.27.8
requests==2.31.0
vocode==0.1.0
python-socketio==5.8.0
elevenlabs==0.2.24
pydub==0.25.1
websockets>=11.0